swarm/storage/mru: Remove redundant test, expose ResourceHash mthd
